Move Identity models to the PortalTeme.Data.Identity namespace.
