+++ PortalTeme.Auth/Areas/Identity/Pages/Account/Manage/ChangePassword.cshtml.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
-using PortalTeme.Data;
+using PortalTeme.Data.Identity;
 namespace PortalTeme.Auth.Areas.Identity.Pages.Account.Manage
 {
     public class ChangePasswordModel : PageModel
