+++ PortalTeme.Auth/Areas/Identity/Pages/_ViewImports.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using PortalTeme.Auth.Areas.Identity
-@using PortalTeme.Data
+@using PortalTeme.Data.Identity
 @namespace PortalTeme.Auth.Areas.Identity.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
