+++ PortalTeme.Auth/Pages/_ViewImports.cshtml
@@ -1,4 +1,4 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using PortalTeme.Data;
+@using PortalTeme.Data.Identity;
 @namespace PortalTeme.Auth.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
