+++ PortalTeme.Data/Identity/User.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using System;
 
-namespace PortalTeme.Data {
+namespace PortalTeme.Data.Identity {
 
     public class User : IdentityUser {
