+++ PortalTeme/appsettings.json
@@ -4,5 +4,8 @@
       "Default": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "PortalTemeContextConnection": "Server=.;Database=portal_teme;Trusted_Connection=True;MultipleActiveResultSets=true"
+  }
 }
